+++ constants.py
@@ -1,4 +1,4 @@
 APP_NAME = "KönyvTárnok"
-APP_VERSION = "0.22.2"
+APP_VERSION = "0.23.0"
 APP_STAGE = "Béta"
 APP_TITLE = f"{APP_NAME} (v{APP_VERSION}) {APP_STAGE}"